#237 Refactor / 모듈 멤버 BurstInterval 자료형 int에서 float으로 변경

+++ Project/Engine/struct.fx
@@ -59,7 +59,7 @@ struct tParticleModule
     float4 vSpawnMaxScale; // 초기 최대 크기
     
     int SpawnType; // 0 : Continuous, 1 : Burst
-    int BurstInterval; // Burst 간격(초 단위)
+    float BurstInterval; // Burst 간격(초 단위)
     float MinLife; // 최소 수명
     float MaxLife; // 최대 수명
     float MinMass; // 최소 질량
